ANDKR-1920 | 고메 검색 결과 > 목록정보에 현 위치에서 떨어진 거리 노출 (0.1km 단위로 표시되지 않음)

+++ dailyHOTEL/app/src/main/java/com/daily/dailyhotel/repository/remote/model/GourmetData.java
@@ -74,7 +74,7 @@ public class GourmetData
     public boolean isSoldOut;
 
     @JsonField(name = "distance")
-    public int distance;
+    public double distance;
 
     @JsonField(name = "addrSummary")
     public String addrSummary;
